Made blocks occupy whole width of screen

+++ include/Displayed/Blocks/Row.hpp
@@ -10,7 +10,7 @@ public:
     Row(Uint8 numberOfBlocks, Uint16 startPosX, Uint16 startPosY, Uint16 blockH,
         Uint8 blockSpacing, Uint16 *windowWidth,
         Uint8 red, Uint8 green, Uint8 blue);
-    Uint8 calculateBlockWidth(Uint16 windowWidth, Uint8 numberOfBlocks,
+    Uint16 calculateBlockWidth(Uint16 windowWidth, Uint8 numberOfBlocks,
                               Uint16 startPosX, Uint8 blockSpacingX);
     std::vector<Block> elements;
